+++ packages/pangea-sdk/src/test/java/cloud/pangeacyber/pangea/audit/ITAuditTest.java
@@ -8,6 +8,7 @@ import static org.junit.Assert.assertTrue;
 import cloud.pangeacyber.pangea.AttachedFile;
 import cloud.pangeacyber.pangea.Config;
 import cloud.pangeacyber.pangea.Helper;
+import cloud.pangeacyber.pangea.Response;
 import cloud.pangeacyber.pangea.ResponseStatus;
 import cloud.pangeacyber.pangea.TestEnvironment;
 import cloud.pangeacyber.pangea.audit.models.*;
@@ -16,6 +17,7 @@ import cloud.pangeacyber.pangea.audit.responses.*;
 import cloud.pangeacyber.pangea.audit.results.LogBulkResult;
 import cloud.pangeacyber.pangea.audit.results.RootResult;
 import cloud.pangeacyber.pangea.exceptions.*;
+import com.fasterxml.jackson.core.type.TypeReference;
 import java.nio.file.Path;
 import java.time.Instant;
 import java.util.LinkedHashMap;
@@ -1139,5 +1141,42 @@ public class ITAuditTest {
 		// Test log stream.
 		final var response = client.logStream(input);
 		assertEquals(ResponseStatus.SUCCESS.toString(), response.getStatus());
+	}
+
+	@Test
+	public void testExportDownload() throws PangeaAPIException, PangeaException {
+		var exportResponse = clientGeneral.export(ExportRequest.builder().end(Instant.now()).verbose(true).build());
+		assertEquals(ResponseStatus.ACCEPTED.toString(), exportResponse.getStatus());
+
+		// Note that the export can easily take dozens of minutes, if not
+		// longer, so we don't actually wait for the results on CI. Instead we
+		// just poll it once and then attempt the download, even when we know it
+		// isn't ready yet, just to verify that the core of the functions are
+		// working.
+		try {
+			clientGeneral.pollResult(exportResponse.getRequestId(), new TypeReference<Response<Void>>() {});
+		} catch (PangeaAPIException error) {
+			assertEquals(ResponseStatus.ACCEPTED.toString(), error.getResponse().getStatus());
+		}
+
+		try {
+			var downloadRequest = new DownloadRequest.Builder().requestId(exportResponse.getRequestId()).build();
+			clientGeneral.downloadResults(downloadRequest);
+		} catch (PangeaAPIException error) {
+			var status = error.getResponse().getStatus();
+
+			// NOT_FOUND is possible if this test runs while an export is
+			// already in progress, since the request ID from the current run
+			// does not match the original request ID that started the export in
+			// a previous run.
+			if (
+				status.equalsIgnoreCase(ResponseStatus.ACCEPTED.toString()) &&
+				status.equalsIgnoreCase(ResponseStatus.NOT_FOUND.toString())
+			) {
+				throw error;
+			}
+		} catch (Exception error) {
+			throw error;
+		}
 	}
 }